fix: address 8 regression and robustness issues across codebase
This commit fixes multiple bugs and adds robustness improvements:
1. Add default_value("set01") to lastz preset CLI args in sd/cross and sd/search
2. Fix PSL span calculation overflow with i64 arithmetic
3. Fix sd run preset argument expansion to pass as separate args
4. Fix sd cluster interval dedup to include strand and species
5. Add negative span validation for LAV file parsing
6. Fix LAV d stanza boundary check off-by-one error
7. Add bounds checking and safe allocation for PGI/HV file parsing
8. Fix tube sort key diagonal offset to handle deeply negative diagonals

All changes include regression tests, pass full test suite, and comply with clippy lint rules.

+++ src/libs/pl/repeat.rs
@@ -291,7 +291,9 @@ pub fn run_align_repeat_pipeline(opts: &AlignRepeatOpts) -> anyhow::Result<()> {
         };
         // Guard against a malformed record with t_end < t_start (a negative
         // difference would wrap into a huge span and pass the length filter).
-        let span = (psl.t_end - psl.t_start).max(0) as usize;
+        // i64 arithmetic: extreme PSL coordinates would overflow the i32
+        // subtraction (e.g. t_start = i32::MIN, t_end = i32::MAX).
+        let span = (psl.t_end as i64 - psl.t_start as i64).max(0) as usize;
         if (psl.ident() as f64) < opts.min_identity || span < opts.min_len {
             continue;
         }
